requirements.txt
dapr==1.8.3
Faker==15.2.0
fastapi==0.86.0
motor==3.0.0
requests==2.28.0
uvicorn==0.19.0
pytz==2022.1
pycryptodome==3.14.1
pycryptodomex==3.14.1
protobuf
dapr-ext-fastapi
python-dotenv
gevent
gunicorn
ujson


